Added a wishlist document.
Modified client to make easier to add connection events.

Fixed crash on server when a player not logged sent a start message.

Added precondition on dumb_client so it doesn't try to start when not
allowed to log.

Removed network_wait constant (not useful outside GTK)

Some random notes added to documentation.